Fixes warnings, removes trailing whitespace, adds @Override annotations.

+++ src/net/java/sip/communicator/impl/metahistory/MetaHistoryServiceImpl.java
@@ -943,7 +943,7 @@ public class MetaHistoryServiceImpl
                 endDate,
                 keywords);
 
-            double part1 = (double)origProgress/
+            double part1 = origProgress/
                 ((double)maxVal*count);
             double convProgress =
                 part1*HistorySearchProgressListener.PROGRESS_MAXIMUM_VALUE +
